+++ test/unit/authentication/auth_workload_identity/attestation_gcp_test.ts
@@ -1,6 +1,6 @@
 import sinon from 'sinon';
 import assert from 'assert';
-import { GoogleAuth } from 'google-auth-library';
+import { GoogleAuth, Impersonated } from 'google-auth-library';
 import {
   getGcpAttestationToken,
   SNOWFLAKE_AUDIENCE,
@@ -8,38 +8,65 @@ import {
 
 describe('Attestation GCP', () => {
   const sinonSandbox = sinon.createSandbox();
-  let getIdTokenClientStub: sinon.SinonStub;
-  let fetchIdTokenStub: sinon.SinonStub;
+  const gcpSdkMock = {
+    getIdTokenClient: sinonSandbox.stub(),
+    getClient: sinonSandbox.stub(),
+    fetchIdToken: sinonSandbox.stub(),
+    impersonatedFetchIdToken: sinonSandbox.stub(),
+  };
 
   beforeEach(async () => {
-    fetchIdTokenStub = sinon.stub();
-    getIdTokenClientStub = sinon.stub().returns({
+    gcpSdkMock.getIdTokenClient.returns({
       idTokenProvider: {
-        fetchIdToken: fetchIdTokenStub,
+        fetchIdToken: gcpSdkMock.fetchIdToken,
       },
     });
-    sinonSandbox.stub(GoogleAuth.prototype, 'getIdTokenClient').callsFake(getIdTokenClientStub);
+    gcpSdkMock.getClient.resolves({});
+    sinonSandbox
+      .stub(GoogleAuth.prototype, 'getIdTokenClient')
+      .callsFake(gcpSdkMock.getIdTokenClient);
+    sinonSandbox.stub(GoogleAuth.prototype, 'getClient').callsFake(gcpSdkMock.getClient);
+    sinonSandbox
+      .stub(Impersonated.prototype, 'fetchIdToken')
+      .callsFake(gcpSdkMock.impersonatedFetchIdToken);
   });
 
   afterEach(() => {
     sinonSandbox.restore();
   });
 
-  it('calls fetchIdToken with the correct audience', async () => {
-    await getGcpAttestationToken();
-    assert.strictEqual(getIdTokenClientStub.firstCall.args[0], SNOWFLAKE_AUDIENCE);
-    assert.strictEqual(fetchIdTokenStub.firstCall.args[0], SNOWFLAKE_AUDIENCE);
+  describe('using default credentials', () => {
+    it('calls fetchIdToken with the correct audience', async () => {
+      await getGcpAttestationToken();
+      assert.strictEqual(gcpSdkMock.getIdTokenClient.firstCall.args[0], SNOWFLAKE_AUDIENCE);
+      assert.strictEqual(gcpSdkMock.fetchIdToken.firstCall.args[0], SNOWFLAKE_AUDIENCE);
+    });
+
+    it('throws error when the token is not found', async () => {
+      const err = new Error('Token not found');
+      gcpSdkMock.fetchIdToken.throws(err);
+      assert.rejects(getGcpAttestationToken(), err);
+    });
+
+    it('returns the token when it is found', async () => {
+      gcpSdkMock.fetchIdToken.resolves('test-token');
+      const token = await getGcpAttestationToken();
+      assert.strictEqual(token, 'test-token');
+    });
   });
 
-  it('throws error when the token is not found', async () => {
-    const err = new Error('Token not found');
-    fetchIdTokenStub.throws(err);
-    assert.rejects(getGcpAttestationToken(), err);
-  });
+  describe('using impersonation', () => {
+    it('throws error when impersonation fails', async () => {
+      const err = new Error('Impersonation failed');
+      gcpSdkMock.impersonatedFetchIdToken.throws(err);
+      await assert.rejects(getGcpAttestationToken(['service-account']), err);
+    });
 
-  it('returns the token when it is found', async () => {
-    fetchIdTokenStub.resolves('test-token');
-    const token = await getGcpAttestationToken();
-    assert.strictEqual(token, 'test-token');
+    it('returns token', async () => {
+      gcpSdkMock.impersonatedFetchIdToken.resolves('impersonated-token');
+      const token = await getGcpAttestationToken(['service-account']);
+      assert.strictEqual(gcpSdkMock.impersonatedFetchIdToken.firstCall.args[0], SNOWFLAKE_AUDIENCE);
+      assert.strictEqual(token, 'impersonated-token');
+    });
   });
 });